requirements.txt
numpy
matplotlib
tensorflow
scikit-learn
sklearn
pyyaml
pandas